Fix rotate touch and change some input values

+++ Assets/Scripts/ActionData.cs
@@ -4,6 +4,6 @@
     {
         public bool isActionHappened;
         public MoveAction action;
-        public float moveDistance;
+        public int moveDistance;
     }
 }
